refactor: move ainize util file.
